+++ src/Highcharts/Highstock.php
@@ -21,6 +21,9 @@ class Highstock extends AbstractChart
 {
     public const CHART_CLASS = 'StockChart';
 
+    /**
+     * Range selector language options for accessibility.
+     */
     public ChartOption $rangeSelector;
 
     public function __construct()